requirements.txt
Package             Version
------------------- -------
asgiref             3.8.1
Django              5.0.4
django-filter       24.2
djangorestframework 3.15.1
Markdown            3.6
pip                 24.0
setuptools          65.5.0
sqlparse            0.5.0
